fix: new observer created on each mutation
