+++ packages/roomkit-react/src/Prebuilt/components/pdfAnnotator/shareScreenOptions.jsx
@@ -28,7 +28,6 @@ export function ShareScreenOptions() {
     });
   };
   const { toggleScreenShare } = useScreenShare();
-
   return (
     <Fragment>
       <Dropdown.Root
